+++ app/Models/Category.php
@@ -4,6 +4,7 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Support\Str;
 
@@ -55,6 +56,11 @@ class Category extends Model
     public function translations(): HasMany
     {
         return $this->hasMany(CategoryTranslation::class);
+    }
+
+    public function licenseTypes(): BelongsToMany
+    {
+        return $this->belongsToMany(LicenseType::class, 'category_license_type');
     }
 
     public function getLocalizedName(?string $locale = null): string